+++ src/extension.ts
@@ -8,11 +8,7 @@ import { activateLsp, deactivateLsp, restartLsp } from "./lsp";
 import { SemgrepDocumentProvider } from "./showAstDocument";
 import { createStatusBar } from "./statusBar";
 import { initTelemetry, stopTelemetry } from "./telemetry/telemetry";
-import {
-  ExtensionEnvironment,
-  topLevelSpan,
-  withSpan,
-} from "./utilities/tracing";
+import { deregisterExistingOtel, withSpan } from "./utilities/tracing";
 import { SemgrepPolicyViewProvider } from "./views/policy";
 import { SemgrepHelpProvider } from "./views/support";
 import { SemgrepSearchWebviewProvider } from "./views/webview";
@@ -116,6 +112,11 @@ async function afterClientStart(context: ExtensionContext, env: Environment) {
 export async function activate(
   context: ExtensionContext,
 ): Promise<Environment | undefined> {
+  // We want to deregister any existing OpenTelemetry global state
+  // as soon as we can, when the language server is started.
+  // See the description of this function for more.
+  deregisterExistingOtel();
+
   const env: Environment = await createOrUpdateEnvironment(context);
   initTelemetry(env);