+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Fixed bug preventing OpenTelemetry traces from being sent from Cursor users
+
 ## 1.12.2 - 2025-07-08
 
 ### Fixed